update: split imputation class from main
